Terminacion de juego, concurrencia musica y partida
